rename all coci files
