--- COCI/3 COCI 06 Contest 2 R2/src/Main.java
@@ -1,11 +1,0 @@
-import java.util.Scanner;
-public class Main {
-	/*
-	 * Copy-pasting code is NOT cool! Please do not copy and paste my code as a submission to DMOJ.
-	 * github.com/jdabtieu/competitive-programming
-	 */
-	static Scanner in = new Scanner(System.in);
-	public static void main(String[] args) {
-		System.out.println(-1 * in.nextInt() + 2 * in.nextInt());
-	}
-}
