+++ COCI/coci06c2p1/src/coci06c2p1.java
@@ -1,0 +1,11 @@
+import java.util.Scanner;
+public class coci06c2p1 {
+	/*
+	 * Copy-pasting code is NOT cool! Please do not copy and paste my code as a submission to DMOJ.
+	 * github.com/jdabtieu/competitive-programming
+	 */
+	static Scanner in = new Scanner(System.in);
+	public static void main(String[] args) {
+		System.out.println(-1 * in.nextInt() + 2 * in.nextInt());
+	}
+}
