+++ COCI/coci08c3p2/src/coci08c3p2.java
@@ -1,0 +1,17 @@
+import java.util.Scanner;
+
+public class coci08c3p2 {
+	/* 
+	 * Copy-pasting code is NOT cool! Please do not copy and paste my code as a submission to DMOJ.
+	 * github.com/jdabtieu/competitive-programming
+	 */
+	public static void main(String[] args) {
+		Scanner in = new Scanner(System.in);
+		String input = in.nextLine();
+		for (byte i = 0; i < input.length() - 2; i++) if (input.charAt(i) == 'a' || input.charAt(i) == 'e' || input.charAt(i) == 'i' || input.charAt(i) == 'o' || input.charAt(i) == 'u') {
+			input = input.substring(0, i + 1) + input.substring(i + 3);
+		}
+		System.out.println(input);
+	}
+
+}
