+++ COCI/coci09c1p2/src/coci09c1p2.java
@@ -1,0 +1,16 @@
+import java.util.Scanner;
+public class coci09c1p2 {
+	/* 
+	 * Copy-pasting code is NOT cool! Please do not copy and paste my code as a submission to DMOJ.
+	 * github.com/jdabtieu/competitive-programming
+	 */
+	static Scanner in = new Scanner(System.in);
+	public static void main(String[] args) {
+		int N = in.nextInt();
+		int sum = 0;
+		for (int i = 0; i <= N; i++) {
+			sum = sum + (N + i) * (N - i + 1) /2 + (N - i + 1) * i;
+		}
+		System.out.println(sum);
+	}
+}
